requirements.txt
torch>=2.0.0
transformers>=4.30.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.1
openai>=1.0.0
gradio>=3.40.0
flask>=2.2.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.3.1
pytest-cov>=4.1.0
flake8>=6.0.0
mypy>=1.5.0
black>=23.3.0
isort>=5.12.0
wandb>=0.15.0
tiktoken>=0.5.0
tweepy>=4.14.0
requests>=2.31.0 